trade_engine/requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
arch==7.2.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==7.1.0
astropy-iers-data==0.2025.6.9.14.9.37
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
binance==0.3.50
bleach==6.2.0
blosc2==3.3.4
bm25s==0.2.13
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
clarabel==0.11.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.11.3
cvxpy==1.6.5
cycler==0.12.1
dask==2025.5.1
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
duckdb==1.3.0
duckdb_engine==0.17.0
ecos==2.0.14
et_xmlfile==2.0.0
eventkit==1.0.3
executing==2.2.0
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.5
fastjsonschema==2.21.1
fonttools==4.58.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.2
fsspec==2025.5.1
future==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperopt==0.2.7
ib-insync==0.9.86
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
llvmlite==0.44.0
lmdb==1.6.2
lmdbm==0.0.6
locket==1.0.0
Mako==1.3.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.18
multitasking==0.0.11
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.10.2
numpy==2.2.6
openai==1.84.0
openpyxl==3.1.5
optuna==4.3.0
osqp==1.0.4
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathos==0.3.4
pathspec==0.12.1
patsy==1.0.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.6.0
pox==0.3.6
ppft==1.7.7
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.3
pyportfolioopt==1.5.6
PyStemmer==3.0.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
QuantStats==0.0.64
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
Riskfolio-Lib==7.0.1
rpds-py==0.25.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.15.3
scs==3.2.7.post2
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
ta-lib @ file:///C:/Users/water/workspace/trade/trade_engine/ta_lib-0.6.3-cp311-cp311-win_amd64.whl#sha256=c3a4e29e7e97dbd03c8d8bef0d26da3ced6cc34cb6af56134d273a0284c748d8
tables==3.10.2
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.2.1
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.34.3
vectorbtpro @ git+https://github.com/polakowo/vectorbt.pro.git@f3bdf4a4b07151380a6e429a08ff53c2a4b3ef56
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
XlsxWriter==3.2.3
yarl==1.20.1
yfinance==0.2.62
zipp==3.23.0
